jtag/drivers: usb_blaster add lowlevel driver libusb-1.0
Add a low-level driver using libusb-1.0.
This is a true asynchronous library, and leverages the most out of the
synchronous to asynchronous conversion.

As the interface is that low level, additionnal blaster protocol must be
done, ie. the status 2 bytes at the beginning of each USB packet sent by
the blaster must be pruned.

Change-Id: I7af550226fa4f524d417e08530496769ee6c148d
Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>